Moved the saving logic into the genotype file for neat
